Main/VideoRecordingAndEncoding: build fix. Current video recording is under #ifdef VBOX_WITH_VPX. Files checked in are still and development and buf fixing and will not be used without flag VBOX_WITH_VPX.
git-svn-id: http://www.virtualbox.org/svn/vbox/trunk@42839 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/include/MachineImpl.h
@@ -421,8 +421,8 @@ public:
     STDMETHOD(COMSETTER(VideoCaptureEnabled))(BOOL  u8VideoRecEnabled);
     STDMETHOD(COMGETTER(VideoCaptureFile))(BSTR * ppChVideoRecFilename);
     STDMETHOD(COMSETTER(VideoCaptureFile))(IN_BSTR pChVideoRecFilename);
-    STDMETHOD(COMGETTER(VideoCaptureWidth))(ULONG *u32VideoRecHorzRes);
-    STDMETHOD(COMSETTER(VideoCaptureWidth))(ULONG u32VideoRecHorzRes);
+    STDMETHOD(COMGETTER(VideoCaptureWidth))(uint32_t *u32VideoRecHorzRes);
+    STDMETHOD(COMSETTER(VideoCaptureWidth))(uint32_t u32VideoRecHorzRes);
     STDMETHOD(COMGETTER(VideoCaptureHeight))(uint32_t *u32VideoRecVertRes);
     STDMETHOD(COMSETTER(VideoCaptureHeight))(uint32_t u32VideoRecVertRes);
     STDMETHOD(COMGETTER(MonitorCount))(ULONG *monitorCount);
